backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
psycopg2-binary
redis
rapidfuzz
torch
transformers
